Working on new structure for shapes
